Adding a function it give safe partition bounds
hour_bound_plus(positive_int)

Starts at the current time, takes the next hour bound and adds
the positive_int o tthat and returns a proper object.

We may need to adjust this as we go forward if we need to support
partitioning other than hourly.  Or we may need to generalize this
and add other supported boundaries.  However this is an initial
start to this problem and can be expanded as our use case expands.

Fixes #34