Procedures to generate servermaps.
Fixes #39

The current approach uses rotation to put the first host on the end and
rotate the serverhosts in this way.  There are a few cases where this
might have unexpected effects.

In the future, we probably need to rotate not by the first host but by
a set number detected through an API.  See #61 for further work