Adding executable program framework
This provides consistent handling of things like the arguments
needed to connect with the Lenkwerk db.  The goal is to ensure
testability in all programs by moving the non-testable components
to a few dedicated places.  A simple program for running the
lenkwerk agent (once that is merged) is included.

Fixes #80