Initial storage agent code.
This includes a main loop and some dispatch table stuff to handle
different types of events coming from the kvstore watches.

It also includes facilities to connect to the storage nodes' databases,
and some fixes to documentation of servermaps (needed for locating
whether an event is relevant to us).

Next I need a generic sink ofr kvstore events into the storage
databases, and then some integration testing to mae sure it all works.
then some of the specialized functionality can be added.