Standardizing tablenames on singular forms
Fixes #92

This is to ease cognitive load of management of database layer.

Note this will to have a clean test run until after the LW Agent
integration test PR #94 is merged due to flaws in the codebase prior
to that PR.

The overall exposure here is low and entirely enforced by Postgres aside
from some minor exposure on the LW Agent side.  This will need to go
through another round of testing after #94 is applied.