Changing "stop" to be more graceful
The AnyEvent::Loop::run() function just does one_event() while 1.

So in this case it is far better to allow a stop() routine to set a stop
variable which is checked in between event polling cycles.  To be safe,
we then set a condvar to send to make sure we are not waiting forever.

The fact that suddenly more tests got run suggests that this has
actually uncovered a bug.