Bumping PGObject::Type::DateTime version to 2.1.0
And removing code that has been merged into that module.

Note that this means that the actual datetime values cannot be used
for the calculating the hour expiration date, that this has to be done
via the package method itself.  However, that's clearer anyway code-wise
and all current code already does things properly.

Fixes #40.