Added notes regarding @_ without unpacking
In general @_ (the default, or argument list of a function) should
always be unpacked before use.  However there are two cases where it can
be clearer/better not to do this.  This commit addresses both of these.

The first is in the case of one-liners.  And the second is in the case
of argument passthrough.