Refactoring to move Schaufel control to own PM
This release makes a few changes to the handling of the processes as
well.  This ensures that the Schaufel object is tied to the lifetime
generally of the process and that if Schaufel dies unexpectedly, the
agent also dies.

The agent still has to maintain the expected state of the next
schaufel run, but doesn't have to generate arguments or the like.

Process exit is now supported with sigterm which gracefully exits the
event loop and terminates Schaufel and sigint, which kills schaufel and
just exits, leaving the OS to clean up things.

We now have a basic retention handle from the agent but it does not
propery work yet.

Next step are integration and unit tests, and then this is ready for
review.