Removed TestDecoder parser and moved that to CPAN
This removes one more problem from the scope of Bagger and ensures that
our code is smaller, more maintainable, etc.

A new problem was encountered along the way which is that t/51-lwagent
is returning a variable number of test results, though these are
certainly being executed in order.  My guess is that some coro is not
finished running when the test exits.  This deserves more investigation.

A final finding is that the undef of the guard for Etcd interferes with
plan parsing and must come after the plan is sent.  This is corrected
here.